Add System Info+Choices

+++ getoj.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 
-Green_font_prefix="\033[32m" && Red_font_prefix="\033[31m" && Green_background_prefix="\033[42;37m" && Red_background_prefix="\033[41;37m" && Font_color_suffix="\033[0m"
+Green_font_prefix="\033[32m" && Red_font_prefix="\033[31m" && Grey_font_preffix="\e[37m" && Green_background_prefix="\033[42;37m" && Red_background_prefix="\033[41;37m" && Font_color_suffix="\033[0m"
+column_size=${COLUMNS}
+length_size=${LINES}
+debug_mode=false
 
 check_root(){
   [[ $EUID != 0 ]] && echo -e "Please use root account or use command 'sudo' to get root access." && exit 1
@@ -12,6 +15,120 @@ print_info(){
 
 print_err(){
   echo -e "${Red_font_prefix}[ERR]$1${Font_color_suffix}"
+}
+
+print_center(){
+  len1=${#1}
+  spaces=`expr 38 - $len1 / 2 `
+  cnt=0
+  while (($cnt<=$spaces))
+  do
+    if [ ! $2 == ' ' ]; then
+       echo -n $2
+    else
+       echo -n ' '
+    fi
+    let "cnt++"
+  done
+  echo -n -e "$1"
+  cnt=0
+  while (($cnt<=$spaces))
+  do
+    if [ ! $2 == ' ' ]; then
+       echo -n $2
+    else
+       echo -n ' '
+    fi
+    let "cnt++"  
+  done
+  printf '\n'
+}
+
+print_grey(){
+  echo -n -e "${Grey_font_size}:${Font_color_suffix}"
+}
+
+get_sysinfo(){
+  cnt=0
+  while (($cnt<80))
+  do
+    echo -n '='
+    let "cnt++"
+  done
+
+  CPU=$(grep 'model name' /proc/cpuinfo |uniq |awk -F : '{print $2}' |sed 's/^[ \t]*//g' |sed 's/ \+/ /g') 
+  printf "%-30s" "CPU Model"
+  print_grey
+  echo -e " ${Green_font_prefix}${CPU}${Font_color_suffix}"
+
+  cpu_num=`cat /proc/cpuinfo|grep 'processor'|sort|uniq|wc -l`
+  printf "%-30s" "CPU Numbers"
+  print_grey
+  echo -e " ${Green_font_prefix}$cpu_num${Font_color_suffix}"
+
+  PROCESSOR=$(grep 'processor' /proc/cpuinfo |sort |uniq |wc -l)
+  printf "%-30s" "Logical CPU Number"
+  print_grey
+  echo -e " ${Green_font_prefix}${PROCESSOR}${Font_color_suffix}"
+
+  Mode=$(getconf LONG_BIT) 
+  printf "%-30s" "CPU Running Mode"
+  print_grey
+  echo -e " ${Green_font_prefix}${Mode}Bits${Font_color_suffix}"
+  
+  Cores=$(grep 'cpu cores' /proc/cpuinfo |uniq |awk -F : '{print $2}' |sed 's/^[ \t]*//g')
+  printf "%-30s" "CPU Cores"
+  print_grey
+  echo -e " ${Green_font_prefix}${Cores}${Font_color_suffix}"
+
+  Total=$(cat /proc/meminfo |grep 'MemTotal' |awk -F : '{print $2}' |sed 's/^[ \t]*//g')
+  printf "%-30s" "Memory in Total"
+  print_grey
+  echo -e " ${Green_font_prefix}${Total}${Font_color_suffix}"
+
+  Available=$(free -m |grep - |awk -F : '{print $2}' |awk '{print $2}')
+  printf "%-30s" "Memory Free"
+  print_grey
+  echo -e " ${Green_font_prefix}${Available}${Font_color_suffix}"
+
+  SwapTotal=$(cat /proc/meminfo |grep 'SwapTotal' |awk -F : '{print $2}' |sed 's/^[ \t]*//g')
+  printf "%-30s" "Swap in Total"
+  print_grey
+  echo -e " ${Green_font_prefix}${SwapTotal}${Font_color_suffix}"
+
+  disk_size=`df -h / | awk '{print $2}'|grep -E '[0-9]'`
+  printf "%-30s" "Disk Size Free"
+  print_grey
+  echo -e " ${Green_font_prefix}${disk_size}${Font_color_suffix}"
+
+  linux_v=$(cat /etc/os-release|grep 'PRETTY_NAME'|cut -c14-)
+  linux_bit=`uname -i`
+  printf "%-30s" "System"
+  print_grey
+  echo -e " ${Green_font_prefix}${linux_v:0:${#linux_v}-1} $linux_bit${Font_color_suffix}"
+
+  process=`ps aux|wc -l`
+  let process--
+  printf "%-30s" "Running Processes"
+  print_grey
+  echo -e " ${Green_font_prefix}$process${Font_color_suffix}"
+
+  software_num=`dpkg -l |wc -l`
+  printf "%-30s" "Software Installed"
+  print_grey
+  echo -e " ${Green_font_prefix}$software_num${Font_color_suffix}"
+
+  kernel_version=$(uname -r)
+  printf "%-30s" "Kernel Version"
+  print_grey
+  echo -e " ${Green_font_prefix}$kernel_version${Font_color_suffix}"
+  cnt=0
+
+  while (($cnt<80))
+  do
+    echo -n '='
+    let "cnt++"
+  done
 }
 
 update_com(){
@@ -32,7 +149,6 @@ install_com(){
 }
 
 check_com(){
-  update_com
   for nowc in 'wget' 'tar' 'unzip' 'python3'
   do
     if command -v ${nowc} >/dev/null 2>&1; then
@@ -80,18 +196,45 @@ install_tomcat(){
 
 install_webapp(){
   hhsoj_ver=$(wget --no-check-certificate -qO- https://api.github.com/repos/XiaoGeNintendo/HHSOJ-Web-Edition/releases | grep -o '"tag_name": ".*"' |head -n 1| sed 's/"//g;s/v//g' | sed 's/tag_name: //g')
-  down_link="https://github.com/XiaoGeNintendo/HHSOJ-Web-Edition/releases/download/${hhsoj_ver}/HellOJ.war"
-  wget -P '/usr/tomcat/webapps/ROOT.war' ${down_link}
-  rm -rf '/usr/tomcat/webapps/ROOT/'
-  print_info "Please now run tomcat to unpack the file" 
+  print_info "Latest WebApp Version:${hhsoj_ver}"
+  read -e -p "Install/Update Now?[Y/n]:" ch
+  if [[  -z $ch ]]; then
+     down_link="https://github.com/XiaoGeNintendo/HHSOJ-Web-Edition/releases/download/${hhsoj_ver}/HellOJ.war"
+     wget -P '/usr/tomcat/webapps/ROOT.war' ${down_link}
+     rm -rf '/usr/tomcat/webapps/ROOT/'
+     print_info "Please now run tomcat to unpack the file"
+  else 
+    if [ "${ch}" == 'y' -o "${ch}" == "Y" ]; then
+      down_link="https://github.com/XiaoGeNintendo/HHSOJ-Web-Edition/releases/download/${hhsoj_ver}/HellOJ.war"
+      wget -P '/usr/tomcat/webapps/ROOT.war' ${down_link}
+      rm -rf '/usr/tomcat/webapps/ROOT/'
+      print_info "Please now run tomcat to unpack the file" 
+    else
+      print_info "Installation/Update Canceled."
+    fi
+  fi
 }
 
 install_folder(){
   down_link=$(wget --no-check-certificate -qO- https://api.github.com/repos/XiaoGeNintendo/HHSOJ-Web-Edition/releases | grep -o 'https://github.com/XiaoGeNintendo/HHSOJ-Web-Edition/releases/download/.*/hhsoj.zip' | head -n 1)
-  wget -P  ${down_link} /usr/hhsoj.zip
-  rm -rf /usr/hhsoj/
-  unzip /usr/hhsoj.zip -d /usr/
-  rm -f /usr/hhsoj.zip
+  folder_ver=$(echo ${down_link} | grep -P '\d+\.\d+' -o)
+  print_info "Latest HHSOJ Folder Version:${folder_ver}"
+  read -e -p "Install/Update Now?[Y/n]:" ch
+  if [[  -z $ch ]]; then
+    wget -P  ${down_link} /usr/hhsoj.zip
+    rm -rf /usr/hhsoj/
+    unzip /usr/hhsoj.zip -d /usr/
+    rm -f /usr/hhsoj.zip
+  else
+   if [ "${ch}" == 'y' -o "${ch}" == "Y" ]; then
+      wget -P  ${down_link} /usr/hhsoj.zip
+      rm -rf /usr/hhsoj/
+      unzip /usr/hhsoj.zip -d /usr/
+      rm -f /usr/hhsoj.zip
+   else
+     print_info "Installation/Update Canceled."
+   fi
+  fi
 }
 
 
@@ -134,14 +277,31 @@ check_all(){
 }
 
 if [ "$#" -gt 0 -a "$1" = "--debug" ]; then
-  if [ ! -f '/usr/tomcat/webapps/ROOT.war/' ]; then
-     print_info 'Found.'
-  else
-     print_err 'Not Found'
-  fi
+  debug_mode=true
+  get_sysinfo
 else
-  check_root
-  check_com
-  check_pip
-  check_all
+  echo -e "${Green_font_prefix}==============================HHSOJ Control Shell===============================${Font_color_suffix}"
+  echo -e "${Green_font_prefix}|                                  By XIZCM                                    |${Font_color_suffix}"
+  echo -e "${Green_font_prefix}|                           HellHoleStudios©, 2019                             |${Font_color_suffix}"
+  echo -e "${Green_font_prefix}================================================================================${Font_color_suffix}"
+  echo ""
+  echo "Operations:"
+  echo "[1]Check&Install HHSOJ"
+  echo "[2]Update HHSOJ"
+  read -e -p "Input Your Choice:" ch
+  case "$ch" in
+    1)
+    check_root
+    update_com
+    check_com
+    check_pip
+    check_all
+    ;;
+    2)
+    print_err 'Under Development'
+    ;;
+    *)
+    print_err 'Please input the right number'
+    ;;
+  esac
 fi
